Added iFrame. Moved mdx styles to separate file.
